requirements.txt
flask==2.0.1
mysql-connector-python==8.0.26
Werkzeug==2.0.1
pandas==1.3.3
xlsxwriter==3.0.2
secrets
datetime
